rename imported modules
